Remove list-plugins file
phovea/generator-phovea#425
